updated tests

+++ include/aare/CalculateEta.hpp
@@ -286,17 +286,20 @@ Eta2<T> calculate_full_eta2(
 }
 
 template <typename T>
-Eta2<T> calculate_eta2(const Cluster<T, 2, 2, int16_t> &cl) {
+Eta2<T> calculate_eta2(const Cluster<T, 2, 2, uint16_t> &cl) {
     Eta2<T> eta{};
 
     if ((cl.data[2] + cl.data[3]) != 0)
         eta.x = static_cast<double>(cl.data[3]) /
-                (cl.data[2] + cl.data[3]); // between (0,1) the closer to zero
-                                           // left value probably larger
+                static_cast<double>(
+                    cl.data[2] + cl.data[3]); // between (0,1) the closer to
+                                              // zero left value probably larger
     if ((cl.data[1] + cl.data[3]) != 0)
-        eta.y = static_cast<double>(cl.data[3]) /
-                (cl.data[1] + cl.data[3]); // between (0,1) the closer to zero
-                                           // bottom value probably larger
+        eta.y =
+            static_cast<double>(cl.data[3]) /
+            static_cast<double>(cl.data[1] +
+                                cl.data[3]); // between (0,1) the closer to zero
+                                             // bottom value probably larger
     eta.sum = cl.sum();
 
     return eta;
